hnn_hm_june_release/requirements.txt
tensorflow-gpu==2.3
matplotlib==3.3.2
scikit-learn==0.23.2
scikit-image==0.17.1
scipy==1.4.1
pybind11==2.4.3
pytest==6.1.1
opencv-contrib-python==4.4.0.46
gdown==3.12.2
pandas==1.0.5
networkx==2.5
pyradiomics==2.24
